Move EditorModel.init to actions.ts::initEditorModel
